+++ pkg/providers/property/testdata/TestDataPropertyHostnameActivations/missing_property_id.tf
@@ -1,0 +1,8 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_property_hostname_activations" "activation" {
+  contract_id = "ctr_1"
+  group_id    = "grp_1"
+}