+++ pkg/providers/property/testdata/TestDataPropertyHostnameActivations/valid_production.tf
@@ -1,0 +1,10 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+data "akamai_property_hostname_activations" "activation" {
+  contract_id = "ctr_1"
+  group_id    = "grp_1"
+  property_id = "prp_1"
+  network     = "PRODUCTION"
+}